refactor(chess/domain): Score gamestate 패키지로 이동
gamestate 패키지는 기본적으로 상태 객체를 담는 패키지지만 다음과 같은 이유로 Score를 포함시킴
    - Score의 비즈니스 메서드를 State 객체들이 사용
    - 점수도 개념적으로 game의 현 상태라고 볼 수 있음
