+++ src/main/java/chess/domain/ChessGame.java
@@ -3,6 +3,7 @@ package chess.domain;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 import chess.domain.gamestate.Ready;
+import chess.domain.gamestate.Score;
 import chess.domain.gamestate.State;
 import java.util.Map;
