+++ src/main/java/chess/domain/gamestate/Running.java
@@ -1,7 +1,6 @@
 package chess.domain.gamestate;
 
 import chess.domain.Camp;
-import chess.domain.Score;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 import java.util.Map;
